- Move `llm_task` and `llm_utils` into separate files - Add `task_id` to `LlmTaskInput`
